[#59] refactor: UserCategory 컴포넌트 분리
